Created database
