+++ python/oop/tasks/2023-07-10.Govno.json
